refactor(runtime): select network dispatcher by host
