+++ packages/script/src/runtime/server/utils/network-dispatcher.platform.ts
@@ -1,0 +1,6 @@
+import type { CreateNetworkDispatcher } from './network-host'
+
+export const createNetworkDispatcher: CreateNetworkDispatcher = async () => ({
+  fetch: globalThis.fetch,
+  close: async () => {},
+})
