+++ vitest.config.ts
@@ -28,6 +28,7 @@ export default defineConfig({
             'unhead/scripts': new URL('./packages/script/node_modules/unhead/dist/scripts.mjs', import.meta.url).pathname,
             '#nuxt-scripts/h3': 'h3',
             '#nuxt-scripts/nitro': new URL('./test/unit/__mocks__/empty.ts', import.meta.url).pathname,
+            '#nuxt-scripts/network-dispatcher': new URL('./packages/script/src/runtime/server/utils/network-dispatcher.node.ts', import.meta.url).pathname,
             // Virtual emitted by the Nuxt module at build time; unit tests
             // mock it via `vi.mock('#build/nuxt-scripts-snippets')`, but the
             // import must first resolve to *something* the bundler accepts.
